Typescriptify compiler
