+++ tsconfig.json
@@ -5,6 +5,7 @@
     "module": "ESNext",
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "skipLibCheck": true,
+    "alwaysStrict": true,
 
     /* Bundler mode */
     "moduleResolution": "bundler",
